+++ packages/gov-ui-kit/src/modules/components/proposal/proposalDataListItem/proposalDataListItemStatus/proposalDataListItemStatus.test.tsx
@@ -2,7 +2,6 @@ import { render, screen } from '@testing-library/react';
 import { DateFormat, IconType, formatterUtils } from '../../../../../core';
 import { modulesCopy } from '../../../../assets';
 import { ProposalStatus } from '../../proposalUtils';
-import { proposalDataListItemUtils } from '../proposalDataListItemUtils';
 import { ProposalDataListItemStatus, type IProposalDataListItemStatusProps } from './proposalDataListItemStatus';
 
 describe('<ProposalDataListItemStatus /> component', () => {
@@ -50,32 +49,50 @@ describe('<ProposalDataListItemStatus /> component', () => {
         expect(screen.queryByTestId(IconType.CALENDAR)).not.toBeInTheDocument();
     });
 
-    test.each(proposalDataListItemUtils.ongoingStatuses)(
-        'displays the date and a pinging indicator when the status is %s and voted is false',
-        (status) => {
-            const date = 1719563030308;
-            render(createTestComponent({ date, status, voted: false }));
+    it('displays the date and a pinging indicator when the status is active and voted is false', () => {
+        const status = ProposalStatus.ACTIVE;
+        const date = 1719563030308;
+        render(createTestComponent({ date, status, voted: false }));
 
-            const formattedDate = formatterUtils.formatDate(date, { format: DateFormat.RELATIVE })!;
-            expect(screen.getByText(formattedDate)).toBeInTheDocument();
-            expect(screen.getByTestId('statePingAnimation')).toBeInTheDocument();
-        },
-    );
+        const formattedDate = formatterUtils.formatDate(date, { format: DateFormat.RELATIVE })!;
+        expect(screen.getByText(formattedDate)).toBeInTheDocument();
+        expect(screen.getByTestId('statePingAnimation')).toBeInTheDocument();
+    });
 
-    test.each(proposalDataListItemUtils.ongoingStatuses)(
-        'displays a you-voted label with an icon checkmark when the status is %s and voted is true',
-        (status) => {
-            render(createTestComponent({ status, voted: true }));
+    it('displays a voted label when the status is active and voted is true', () => {
+        const status = ProposalStatus.ACTIVE;
+        render(createTestComponent({ status, voted: true }));
 
-            expect(screen.getByText(/You've voted/i)).toBeInTheDocument();
-            expect(screen.getByTestId(IconType.CHECKMARK)).toBeInTheDocument();
-        },
-    );
+        expect(screen.getByText(/Voted/i)).toBeInTheDocument();
+        expect(screen.getByTestId(IconType.CHECKMARK)).toBeInTheDocument();
+    });
 
-    it('does not display a you-voted label when the status is not an ongoing one and the voted is true', () => {
+    it('does not display a voted label when the status is not active and voted is true', () => {
         render(createTestComponent({ status: ProposalStatus.EXECUTED, voted: true }));
 
-        expect(screen.queryByText(/You've voted/i)).not.toBeInTheDocument();
+        expect(screen.queryByText(/Voted/i)).not.toBeInTheDocument();
         expect(screen.queryByTestId(IconType.CHECKMARK)).not.toBeInTheDocument();
+    });
+
+    it('displays the status context when statusContext is provided and status is active', () => {
+        const statusContext = 'Stage 1';
+        const status = ProposalStatus.ACTIVE;
+        render(createTestComponent({ statusContext, status }));
+        expect(screen.getByText(statusContext)).toBeInTheDocument();
+    });
+
+    it('displays the status context when statusContext is provided and status is advanceable', () => {
+        const statusContext = 'Stage 1';
+        const status = ProposalStatus.ADVANCEABLE;
+        render(createTestComponent({ status, statusContext }));
+        expect(screen.getByText(modulesCopy.proposalDataListItemStatus.statusLabel[status])).toBeInTheDocument();
+        expect(screen.getByText(statusContext)).toBeInTheDocument();
+    });
+
+    it('does not display the status context when statusContext is provided and status is final', () => {
+        const statusContext = 'Stage 1';
+        const status = ProposalStatus.EXECUTED;
+        render(createTestComponent({ statusContext, status }));
+        expect(screen.queryByText(statusContext)).not.toBeInTheDocument();
     });
 });