+++ packages/gov-ui-kit/src/modules/components/proposal/proposalDataListItem/proposalDataListItemStatus/proposalDataListItemStatus.tsx
@@ -2,27 +2,19 @@ import classNames from 'classnames';
 import {
     AvatarIcon,
     DateFormat,
+    formatterUtils,
     IconType,
     Rerender,
     StatePingAnimation,
     Tag,
-    formatterUtils,
-    type StatePingAnimationVariant,
 } from '../../../../../core';
 import type { ModulesCopy } from '../../../../assets';
 import { useGukModulesContext } from '../../../gukModulesProvider';
 import { ProposalStatus, proposalStatusToTagVariant } from '../../proposalUtils';
 import { type IProposalDataListItemStructureProps } from '../proposalDataListItemStructure';
-import { proposalDataListItemUtils } from '../proposalDataListItemUtils';
 
 export interface IProposalDataListItemStatusProps
-    extends Pick<IProposalDataListItemStructureProps, 'date' | 'status' | 'voted'> {}
-
-const proposalStatusToPingVariant = new Map<ProposalStatus, StatePingAnimationVariant>([
-    [ProposalStatus.ACTIVE, 'info'],
-    [ProposalStatus.CHALLENGED, 'warning'],
-    [ProposalStatus.VETOED, 'warning'],
-]);
+    extends Pick<IProposalDataListItemStructureProps, 'date' | 'status' | 'voted' | 'statusContext'> {}
 
 const getFormattedProposalDate = (date: string | number, now: number, copy: ModulesCopy) => {
     const formattedDuration = formatterUtils.formatDate(date, { format: DateFormat.DURATION })!;
@@ -34,41 +26,46 @@ const getFormattedProposalDate = (date: string | number, now: number, copy: Modu
 };
 
 export const ProposalDataListItemStatus: React.FC<IProposalDataListItemStatusProps> = (props) => {
-    const { date, status, voted } = props;
+    const { date, status, statusContext, voted } = props;
 
-    const isOngoing = proposalDataListItemUtils.isOngoingStatus(status);
-    const isOngoingAndVoted = isOngoing && voted;
+    const isActive = status === ProposalStatus.ACTIVE;
+
+    const showStatusContext = statusContext != null && (isActive || status === ProposalStatus.ADVANCEABLE);
     const showStatusMetadata = status !== ProposalStatus.DRAFT;
 
     const { copy } = useGukModulesContext();
 
     return (
-        <div className="flex w-full items-center justify-between gap-x-4 md:gap-x-6">
-            <Tag
-                label={copy.proposalDataListItemStatus.statusLabel[status]}
-                variant={proposalStatusToTagVariant[status]}
-                className="shrink-0"
-            />
+        <div className="flex min-h-6 w-full items-center justify-between gap-x-4 md:gap-x-6">
+            <div className="flex min-w-0 items-center gap-x-1">
+                <Tag
+                    label={copy.proposalDataListItemStatus.statusLabel[status]}
+                    variant={proposalStatusToTagVariant[status]}
+                    className="shrink-0"
+                />
+                {showStatusContext && (
+                    <div className="truncate text-sm leading-tight md:text-base">
+                        <span className="text-neutral-500">{copy.proposalDataListItemStatus.in} </span>
+                        {statusContext}
+                    </div>
+                )}
+            </div>
             {showStatusMetadata && (
                 <div className="flex items-center gap-x-2 md:gap-x-3">
                     <span
                         className={classNames('text-sm leading-tight md:text-base', {
-                            'text-info-800': status === ProposalStatus.ACTIVE,
-                            'text-warning-800':
-                                status === ProposalStatus.CHALLENGED || status === ProposalStatus.VETOED,
-                            'text-neutral-800': !isOngoing,
+                            'text-info-800': isActive,
+                            'text-neutral-800': !isActive,
                         })}
                     >
-                        {isOngoingAndVoted && copy.proposalDataListItemStatus.voted}
-                        {!isOngoingAndVoted && date != null && (
+                        {isActive && voted && copy.proposalDataListItemStatus.voted}
+                        {(!isActive || !voted) && date != null && (
                             <Rerender>{(now) => getFormattedProposalDate(date, now, copy)}</Rerender>
                         )}
                     </span>
-                    {isOngoingAndVoted && <AvatarIcon icon={IconType.CHECKMARK} responsiveSize={{ md: 'md' }} />}
-                    {isOngoing && !voted && <StatePingAnimation variant={proposalStatusToPingVariant.get(status)} />}
-                    {!isOngoing && !voted && date && (
-                        <AvatarIcon icon={IconType.CALENDAR} responsiveSize={{ md: 'md' }} />
-                    )}
+                    {isActive && voted && <AvatarIcon icon={IconType.CHECKMARK} size="sm" />}
+                    {isActive && !voted && <StatePingAnimation variant="info" />}
+                    {!isActive && date && <AvatarIcon icon={IconType.CALENDAR} size="sm" />}
                 </div>
             )}
         </div>