+++ packages/gov-ui-kit/src/modules/components/proposal/proposalDataListItem/proposalDataListItemStructure/proposalDataListItemStructure.api.ts
@@ -2,15 +2,7 @@ import type { IDataListItemProps } from '../../../../../core';
 import { type ICompositeAddress, type IWeb3ComponentProps } from '../../../../types';
 import { type ProposalStatus } from '../../proposalUtils';
 
-export type ProposalType = 'majorityVoting' | 'approvalThreshold' | undefined;
-
-export type ProposalResult<TType extends ProposalType> = TType extends 'majorityVoting'
-    ? IMajorityVotingResult
-    : TType extends 'approvalThreshold'
-      ? IApprovalThresholdResult
-      : undefined;
-
-export type IProposalDataListItemStructureBaseProps<TType extends ProposalType = ProposalType> = IDataListItemProps &
+export type IProposalDataListItemStructureProps = IDataListItemProps &
     IWeb3ComponentProps & {
         /**
          * Proposal id
@@ -29,13 +21,14 @@ export type IProposalDataListItemStructureBaseProps<TType extends ProposalType =
          */
         publisher: IPublisher | IPublisher[];
         /**
-         * Result of the proposal shown only when it is active, challenged or vetoed.
-         */
-        result?: ProposalResult<TType>;
-        /**
          * Proposal status
          */
         status: ProposalStatus;
+        /**
+         * Provides additional context about the current status of a proposal within a multistage voting process.
+         * Only displayed when proposal status is `ACTIVE` or `ADVANCEABLE`.
+         */
+        statusContext?: string;
         /**
          * Proposal description
          */
@@ -44,10 +37,6 @@ export type IProposalDataListItemStructureBaseProps<TType extends ProposalType =
          * Proposal title
          */
         title: string;
-        /**
-         * Type of the ProposalDataListItem
-         */
-        type?: TType;
         /**
          * Indicates whether the connected wallet has voted
          */
@@ -60,52 +49,3 @@ export interface IPublisher extends ICompositeAddress {
      */
     link?: string;
 }
-
-export interface IProposalStage {
-    /**
-     * Name of the proposal stage
-     */
-    title?: string;
-    /**
-     * Id of the proposal stage
-     */
-    id: string | number;
-}
-
-export interface IProposalResultBase {
-    /**
-     * Proposal stage
-     */
-    stage?: IProposalStage;
-}
-
-export interface IApprovalThresholdResult extends IProposalResultBase {
-    /**
-     * Number of approvals for the proposal
-     */
-    approvalAmount: number;
-    /**
-     * Proposal approval threshold
-     */
-    approvalThreshold: number;
-}
-
-export interface IMajorityVotingResult extends IProposalResultBase {
-    /**
-     * Winning option
-     */
-    option: string;
-    /**
-     * Number of votes for the winning option
-     */
-    voteAmount: string;
-    /**
-     * Percentage of votes for the winning option
-     */
-    votePercentage: number;
-}
-
-export type IProposalDataListItemStructureProps =
-    | IProposalDataListItemStructureBaseProps<undefined>
-    | IProposalDataListItemStructureBaseProps<'majorityVoting'>
-    | IProposalDataListItemStructureBaseProps<'approvalThreshold'>;