+++ packages/gov-ui-kit/src/modules/components/proposal/proposalDataListItem/proposalDataListItemUtils.ts
@@ -1,9 +1,0 @@
-import { ProposalStatus } from '../proposalUtils';
-
-class ProposalDataListItemUtils {
-    ongoingStatuses: ProposalStatus[] = [ProposalStatus.ACTIVE, ProposalStatus.CHALLENGED, ProposalStatus.VETOED];
-
-    isOngoingStatus = (status: ProposalStatus) => this.ongoingStatuses.includes(status);
-}
-
-export const proposalDataListItemUtils = new ProposalDataListItemUtils();