+++ packages/gov-ui-kit/CHANGELOG.md
@@ -11,9 +11,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Update `InputFileAvatar` core component to remove `onFileSelect` property and use `value` and `onChange` properties
   instead
-- Update `IProposalActionUpdateMetadataDaoMetadata` interface `logo` property to `avatar` to better align with actions.
+- Update `IProposalActionUpdateMetadataDaoMetadata` interface `logo` property to `avatar` to better align with actions
 - Bump `softprops/action-gh-release` from 2.2.0 to 2.2.1
 - Update minor and patch NPM dependencies
+- Update `<ProposalDataListItem />` module component to include `statusContext` property and support `ADVANCEABLE`
+  status
+- Remove `<MajorityVotingResult />` and `<ApprovalThresholdResult />` sub-components for `<ProposalDataListItem />`
 
 ## [1.0.62] - 2025-01-08